+++ be1-go/message2/messagedata/lao_create.go
@@ -6,7 +6,7 @@ type LaoCreate struct {
 	Action    string   `json:"action"`
 	ID        string   `json:"id"`
 	Name      string   `json:"name"`
-	Creation  int      `json:"creation"`
+	Creation  int64    `json:"creation"`
 	Organizer string   `json:"organizer"`
 	Witnesses []string `json:"witnesses"`
 }
